Use a package instead of a single module
