+++ github_users/__init__.py
@@ -1,0 +1,3 @@
+from github_users.application import Application
+from github_users.cli import main
+from github_users.__version__ import __version__
