+++ setup.py
@@ -1,28 +1,27 @@
 # -*- coding: utf-8 -*-
 
 from __future__ import absolute_import
-from setuptools import setup
-from github_users import Application
+from setuptools import setup, find_packages
+from github_users import __version__
 
-
-# We use the version to construct the DOWNLOAD_URL.
-VERSION = Application.VERSION
 
 # URL to the repository on Github.
 REPO_URL = 'https://github.com/hanpeter/github-users'
 
 # Github will generate a tarball as long as you tag your releases, so don't
 # forget to tag!
-DOWNLOAD_URL = ''.join((REPO_URL, '/tarball/release/', VERSION))
+DOWNLOAD_URL = ''.join((REPO_URL, '/tarball/release/', __version__))
 
 
 setup(
     name='github-users',
-    version=VERSION,
-    author='Peter Han <git@peterhan.me>',
+    version=__version__,
+    author='Peter Han',
+    author_email='git@peterhan.me',
     description='CLI that returns a list of users in a GitHub organization',
     url=REPO_URL,
     download_url=DOWNLOAD_URL,
+    packages=find_packages(exclude=['tests']),
     classifiers=[
         'Development Status :: 4 - Beta',
         'License :: OSI Approved :: MIT License',
@@ -32,10 +31,9 @@ setup(
         'Topic :: Software Development :: Documentation',
         'Topic :: Utilities',
     ],
-    py_modules=['github_users'],
     install_requires=[
         'click',
-        'github3.py'
+        'github3.py',
     ],
     entry_points={
         'console_scripts': [
